+++ core/src/main/java/org/testifyproject/core/reifier/FakeCollaboratorReifier.java
@@ -15,24 +15,26 @@
  */
 package org.testifyproject.core.reifier;
 
-import java.util.Optional;
 import org.testifyproject.MockProvider;
 import org.testifyproject.TestContext;
 import org.testifyproject.TestDescriptor;
-import org.testifyproject.core.util.ReflectionUtil;
-import org.testifyproject.extension.FieldReifier;
 import org.testifyproject.tools.Discoverable;
+import org.testifyproject.extension.annotation.IntegrationCategory;
+import org.testifyproject.extension.annotation.SystemCategory;
 import org.testifyproject.extension.annotation.UnitCategory;
+import org.testifyproject.extension.CollaboratorReifier;
 
 /**
  * A class that reifies test fields annotated with
- * {@link  org.testifyproject.annotation.Virtual}.
+ * {@link  org.testifyproject.annotation.Fake}.
  *
  * @author saden
  */
 @UnitCategory
+@IntegrationCategory
+@SystemCategory
 @Discoverable
-public class VirtualFieldReifier implements FieldReifier {
+public class FakeCollaboratorReifier implements CollaboratorReifier {
 
     @Override
     public void reify(TestContext testContext) {
@@ -41,22 +43,19 @@ public class VirtualFieldReifier implements FieldReifier {
         MockProvider mockProvider = testContext.getMockProvider();
 
         testDescriptor.getFieldDescriptors().parallelStream()
-                .filter(p -> p.getVirtual().isPresent())
+                .filter(p -> p.getFake().isPresent())
                 .forEach(fieldDescriptor -> {
                     Class<?> fieldType = fieldDescriptor.getType();
-                    Object fieldValue = null;
 
-                    Optional<Object> foundFieldValue = fieldDescriptor.getValue(testInstance);
+                    Object fieldValue = fieldDescriptor.getValue(testInstance)
+                            .map(value -> {
+                                if (mockProvider.isMock(value)) {
+                                    return value;
+                                }
 
-                    if (foundFieldValue.isPresent()) {
-                        fieldValue = foundFieldValue.get();
-                    } else if (!fieldType.isInterface()) {
-                        fieldValue = ReflectionUtil.INSTANCE.newInstance(fieldType);
-                    }
-
-                    if (fieldValue != null && !mockProvider.isMock(fieldValue)) {
-                        fieldValue = mockProvider.createVirtual(fieldType, fieldValue);
-                    }
+                                return mockProvider.createVirtual(fieldType, value);
+                            })
+                            .orElseGet(() -> mockProvider.createFake(fieldType));
 
                     fieldDescriptor.setValue(testInstance, fieldValue);
                 });