+++ level/integration/src/main/java/org/testifyproject/level/integration/IntegrationTestRunner.java
@@ -28,7 +28,6 @@ import org.testifyproject.TestResourcesProvider;
 import org.testifyproject.TestRunner;
 import static org.testifyproject.core.TestContextProperties.SERVICE_INSTANCE;
 import org.testifyproject.core.util.ServiceLocatorUtil;
-import org.testifyproject.extension.FieldReifier;
 import org.testifyproject.extension.FinalReifier;
 import org.testifyproject.extension.InitialReifier;
 import org.testifyproject.extension.PostVerifier;
@@ -36,6 +35,7 @@ import org.testifyproject.extension.PreVerifier;
 import org.testifyproject.extension.PreiVerifier;
 import org.testifyproject.extension.annotation.IntegrationCategory;
 import org.testifyproject.tools.Discoverable;
+import org.testifyproject.extension.CollaboratorReifier;
 
 /**
  * A class used to run a integration test.
@@ -64,7 +64,7 @@ public class IntegrationTestRunner implements TestRunner {
         Optional<SutDescriptor> foundSutDescriptor = testContext.getSutDescriptor();
         TestDescriptor testDescriptor = testContext.getTestDescriptor();
 
-        serviceLocatorUtil.findAllWithFilter(FieldReifier.class, IntegrationCategory.class)
+        serviceLocatorUtil.findAllWithFilter(CollaboratorReifier.class, IntegrationCategory.class)
                 .forEach(p -> p.reify(testContext));
 
         serviceLocatorUtil.findAllWithFilter(PreVerifier.class, testDescriptor.getGuidelines(), IntegrationCategory.class)