+++ level/unit/src/main/java/org/testifyproject/level/unit/UnitTestRunner.java
@@ -21,7 +21,6 @@ import org.testifyproject.TestContext;
 import org.testifyproject.TestDescriptor;
 import org.testifyproject.TestRunner;
 import org.testifyproject.core.util.ServiceLocatorUtil;
-import org.testifyproject.extension.FieldReifier;
 import org.testifyproject.extension.FinalReifier;
 import org.testifyproject.extension.InitialReifier;
 import org.testifyproject.extension.PostVerifier;
@@ -30,6 +29,7 @@ import org.testifyproject.extension.PreiVerifier;
 import org.testifyproject.extension.SutReifier;
 import org.testifyproject.extension.annotation.UnitCategory;
 import org.testifyproject.tools.Discoverable;
+import org.testifyproject.extension.CollaboratorReifier;
 
 /**
  * A class used to run a integration test.
@@ -66,7 +66,7 @@ public class UnitTestRunner implements TestRunner {
                     .forEach(p -> p.reify(testContext));
         }
 
-        serviceLocatorUtil.findAllWithFilter(FieldReifier.class, UnitCategory.class)
+        serviceLocatorUtil.findAllWithFilter(CollaboratorReifier.class, UnitCategory.class)
                 .forEach(p -> p.reify(testContext));
 
         serviceLocatorUtil.findAllWithFilter(FinalReifier.class, UnitCategory.class)